add helper package tests
